requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
faiss-cpu
numpy
PyPDF2
python-docx
chardet
requests
beautifulsoup4
